refactor [James]: refactor admin routes and login
