+++ server-side/routes/admin.js
@@ -1,899 +1,167 @@
 const express = require("express");
-const bcrypt = require("bcryptjs");
-const Admin = require("../models/AdminModel");
-const MembershipHistory = require("../models/MembershipHistory");
-const Merch = require("../models/MerchModel");
-const Student = require("../models/StudentModel");
-const Order = require("../models/OrdersModel");
-const Log = require("../models/LogModel");
-const { format, startOfDay, endOfDay } = require("date-fns");
-const nodemailer = require("nodemailer");
-const ejs = require("ejs");
-const path = require("path");
+const {
+  getSearchStudentByIdController,
+  approveMembershipController,
+  revokeAllMembershipController,
+  getMembershipHistoryController,
+  getMembershipRequestController,
+  getStudentsCountController,
+  getActiveMembershipCountController,
+  getPublishMerchandiseCountController,
+  getOrderPlacedCountController,
+  getStudentDashboardCountController,
+  getDailySalesController,
+  getAllAdminAccountsController,
+  getAllAdminMembersController,
+  getAllSuspendAdminAccountController,
+  editAdminAccountController,
+  changeAdminPasswordController,
+  setSuspendAdminAccountController,
+  setMemberRoleRemoveController,
+  setRestoreAdminAccountController,
+  setAdminRequestRoleController,
+  getAllRequestMemberController,
+  getAllRequestAdminAccountController,
+  approveRoleMemberController,
+  setDeclineMemberRoleController,
+  addNewAdminAccountController,
+  approveAdminAccountController,
+  declineAdminAccountController,
+  setNewAdminAccessController,
+} = require("../controllers/admin.controller");
 const {
   admin_authenticate,
 } = require("../middlewares/custom_authenticate_token");
-const mongoose = require("mongoose");
 
 const router = express.Router();
 
 //Search student by ID
 router.get(
-  "/admin/student_search/:id_number",
+  "/student_search/:id_number",
   admin_authenticate,
-  async (req, res) => {
-    const { id_number } = req.params;
-
-    try {
-      const student = await Student.findOne({
-        id_number,
-      });
-
-      if (!student) {
-        res.status(404).json({
-          message: "Student not found!",
-        });
-      } else {
-        res.status(200).json({ data: student });
-      }
-    } catch (error) {
-      console.error(error);
-      res
-        .status(500)
-        .json({ message: "An error occurred", error: error.message });
-    }
-  }
+  getSearchStudentByIdController
 );
-
-router.post("/approve-membership", admin_authenticate, async (req, res) => {
-  const { reference_code, id_number, type, admin, rfid, date, cash, total } =
-    req.body;
-
-  const session = await mongoose.startSession();
-  session.startTransaction();
-
-  try {
-    const student = await Student.findOne({ id_number }).session(session);
-
-    if (!student) {
-      console.error(`Student with id_number ${id_number} not found.`);
-      return res.status(404).json({ message: "Student not found" });
-    }
-
-    let updateQuery = {};
-    if (student.isFirstApplication) {
-      updateQuery = {
-        membershipStatus: "ACTIVE",
-        rfid: rfid,
-        isFirstApplication: false,
-      };
-    } else {
-      updateQuery = {
-        membershipStatus: "RENEWED",
-      };
-    }
-
-    await Student.updateOne({ id_number }, { $set: updateQuery }).session(
-      session
-    );
-
-    const history = new MembershipHistory({
-      id_number,
-      rfid,
-      reference_code,
-      type: student.isFirstApplication ? "Membership" : "Renewal",
-      name: `${student.first_name} ${student.middle_name} ${student.last_name}`,
-      year: student.year,
-      course: student.course,
-      date: new Date(),
-      admin: admin ? admin : req.user.name,
-    });
-
-    const savedHistory = await history.save();
-
-    if (!savedHistory) {
-      console.error("Failed to save membership history.");
-
-      return res
-        .status(500)
-        .json({ message: "Failed to save membership history" });
-    }
-    await session.commitTransaction();
-    session.endSession();
-
-    const transporter = nodemailer.createTransport({
-      service: "gmail",
-      auth: {
-        user: process.env.EMAIL,
-        pass: process.env.PASSWORD_APP_EMAIL,
-      },
-    });
-
-    const emailTemplate = await ejs.renderFile(
-      path.join(__dirname, "../templates/appr-membership-receipt.ejs"), // Path to the ejs file
-      {
-        name: `${student.first_name} ${student.middle_name} ${student.last_name}`,
-        reference_code,
-        cash: cash ? cash : 50,
-        total: cash ? cash : 50,
-        course: student.course,
-        year: student.year,
-        admin: admin ? admin : req.user.name,
-        date: format(new Date(), "MMMM d, yyyy"),
-        change: cash - total,
-      }
-    );
-
-    const mailOptions = {
-      from: process.env.EMAIL,
-      to: student.email,
-      subject: "Your Receipt from PSITS - UC Main",
-      html: emailTemplate,
-      attachments: [
-        {
-          filename: "psits.jpg",
-          path: path.join(__dirname, "../src/psits.jpg"),
-          cid: "logo",
-        },
-      ],
-    };
-
-    // Send the email
-    transporter.sendMail(mailOptions, (error, info) => {
-      if (error) {
-        console.error("Error sending email:", error);
-      } else {
-        console.log("Email sent: " + info.response);
-      }
-    });
-    return res
-      .status(200)
-      .json({ message: "Membership approved successfully" });
-  } catch (error) {
-    console.error("Internal server error:", error);
-    await session.abortTransaction();
-    session.endSession();
-    res
-      .status(500)
-      .json({ message: "Internal server error", error: error.message });
-  }
-});
-
-router.put("/renew-student", admin_authenticate, async (req, res) => {
-  try {
-    const renewStudent = await Student.updateMany({
-      membershipStatus: "NOT_APPLIED",
-    });
-
-    if (!renewStudent) {
-      return res.status(404).json({ message: "Student not found" });
-    }
-
-    res
-      .status(200)
-      .json({ message: "All Student has been renewed successfully" });
-  } catch (error) {
-    console.error("Error deleting student:", error);
-    res.status(500).json("Internal Server Error");
-  }
-});
-
-router.get("/history", admin_authenticate, async (req, res) => {
-  try {
-    const students = await MembershipHistory.find().sort({ date: -1 });
-    res.status(200).json(students);
-  } catch (error) {
-    console.error("Error fetching students:", error);
-    res.status(500).json("Internal Server Error");
-  }
-});
-
-router.get("/renew", admin_authenticate, async (req, res) => {
-  try {
-    const students = await Student.find({ renew: "Pending" });
-    res.status(200).json(students);
-  } catch (error) {
-    console.error("Error fetching students:", error);
-    res.status(500).json("Internal Server Error");
-  }
-});
-
-router.get("/membershipRequest", admin_authenticate, async (req, res) => {
-  try {
-    const students = await Student.find({ membershipStatus: "PENDING" });
-    res.status(200).json(students);
-  } catch (error) {
-    console.error("Error fetching students:", error);
-    res.status(500).json("Internal Server Error");
-  }
-});
-
-router.get("/get-students-count", admin_authenticate, async (req, res) => {
-  try {
-    const [all, request, renew, deleted, history] = await Promise.all([
-      Student.countDocuments({
-        status: "True",
-      }),
-      Student.countDocuments({ membershipStatus: "PENDING" }),
-      Student.countDocuments({ renew: "Pending" }),
-      Student.countDocuments({ status: "False" }),
-      MembershipHistory.countDocuments(),
-    ]);
-
-    res.status(200).json({ all, request, renew, deleted, history });
-  } catch (error) {
-    console.error(error);
-  }
-});
+//ApproveMembership
+router.post(
+  "/approve-membership",
+  admin_authenticate,
+  approveMembershipController
+);
+//Revoke Student Membership
+router.put(
+  "/revoke-student",
+  admin_authenticate,
+  revokeAllMembershipController
+);
+//Membership History
+router.get("/history", admin_authenticate, getMembershipHistoryController);
+//Membership Request
+router.get(
+  "/membership-request",
+  admin_authenticate,
+  getMembershipRequestController
+);
+//Dashboard Student Counts
+router.get(
+  "/get-students-count",
+  admin_authenticate,
+  getStudentsCountController
+);
 //get-active-membership-count
 router.get(
   "/get-active-membership-count",
   admin_authenticate,
-  async (req, res) => {
-    try {
-      const count = await Student.countDocuments({
-        status: "True",
-        $or: [{ membershipStatus: "ACTIVE" }, { membershipStatus: "RENEWED" }],
-      });
-      res.status(200).json({ message: count });
-    } catch (error) {
-      console.error("Error fetching active membership count:", error);
-      res.status(500).json({ error: "Failed to fetch count" });
-    }
-  }
+  getActiveMembershipCountController
 );
-
-router.get("/merchandise-created", admin_authenticate, async (req, res) => {
-  const now = new Date();
-
-  const count = await Merch.countDocuments({
-    is_active: true,
-    start_date: { $lte: now },
-    end_date: { $gte: now },
-  });
-  return res.json({ message: count });
-});
-router.get("/placed-orders", admin_authenticate, async (req, res) => {
-  const count = await Order.countDocuments({ order_status: "Pending" });
-  return res.json({ message: count });
-});
-router.get("/dashboard-stats", admin_authenticate, async (req, res) => {
-  try {
-    const [
-      bsitCount,
-      bscsCount,
-      actCount,
-      year1Count,
-      year2Count,
-      year3Count,
-      year4Count,
-    ] = await Promise.all([
-      Student.countDocuments({ course: "BSIT" }),
-      Student.countDocuments({ course: "BSCS" }),
-      Student.countDocuments({ course: "ACT" }),
-      Student.countDocuments({ year: "1" }),
-      Student.countDocuments({ year: "2" }),
-      Student.countDocuments({ year: "3" }),
-      Student.countDocuments({ year: "4" }),
-    ]);
-
-    return res.json({
-      courses: {
-        BSIT: bsitCount,
-        BSCS: bscsCount,
-        ACT: actCount,
-      },
-      years: {
-        year1: year1Count,
-        year2: year2Count,
-        year3: year3Count,
-        year4: year4Count,
-      },
-    });
-  } catch (error) {
-    return res
-      .status(500)
-      .json({ error: "An error occurred while fetching statistics." });
-  }
-});
-
-router.get("/get-order-date", admin_authenticate, async (req, res) => {
-  try {
-    const currentDate = new Date();
-    const startOfDayDate = startOfDay(currentDate);
-    const endOfDayDate = endOfDay(currentDate);
-
-    const result = await Order.aggregate([
-      {
-        $match: {
-          transaction_date: {
-            $gte: startOfDayDate,
-            $lte: endOfDayDate,
-          },
-          order_status: "Paid",
-        },
-      },
-      { $unwind: "$items" },
-      {
-        $group: {
-          _id: "$items.product_name",
-          totalQuantity: { $sum: "$items.quantity" },
-          totalSubtotal: { $sum: "$items.sub_total" },
-        },
-      },
-      {
-        $project: {
-          product_name: "$_id",
-          totalQuantity: 1,
-          totalSubtotal: 1,
-          _id: 0,
-        },
-      },
-    ]);
-
-    res.json(result);
-  } catch (error) {
-    console.error("Error fetching orders by date:", error);
-    res.status(500).json({ error: "An error occurred while fetching orders" });
-  }
-});
-
-//get all officers
-router.get("/get-all-officers", async (req, res) => {
-  try {
-    const officers = await Admin.find({ status: "Active" });
-
-    const users = officers.map((officer) => ({
-      id_number: officer.id_number,
-      email: officer.email,
-      name: officer.name,
-      course: officer.course,
-      year: officer.year,
-      position: officer.position,
-      campus: officer.campus,
-      access: officer.access,
-    }));
-
-    res.status(200).json({ data: users });
-  } catch (error) {
-    console.error("Error fetching officers:", error);
-    res.status(500).json({ error: "Failed to fetch officers" });
-  }
-});
-
-//TODO: get all MEMBERS
-
-router.get("/get-all-members", async (req, res) => {
-  try {
-    const rolesToFind = ["developer", "officers", "media", "volunteer"];
-    const members = await Student.find({
-      role: { $in: rolesToFind },
-      isRequest: false,
-    });
-
-    const users = members.map((members) => ({
-      id_number: members.id_number,
-      email: members.email,
-      name:
-        members.first_name +
-        " " +
-        members.middle_name +
-        " " +
-        members.last_name,
-      course: members.course,
-      year: members.year,
-      role: members.role,
-      status: members.status,
-    }));
-
-    res.status(200).json({ data: users });
-  } catch (error) {
-    console.error("Error fetching officers:", error);
-    res.status(500).json({ error: "Failed to fetch officers" });
-  }
-});
-
-router.get("/get-suspend-officers", admin_authenticate, async (req, res) => {
-  try {
-    const officers = await Admin.find({ status: "Suspend" });
-
-    const users = officers.map((officer) => ({
-      id_number: officer.id_number,
-      email: officer.email,
-      name: officer.name,
-      course: officer.course,
-      year: officer.year,
-      position: officer.position,
-    }));
-
-    res.status(200).json({ data: users });
-  } catch (error) {
-    console.error("Error fetching officers:", error);
-    res.status(500).json({ error: "Failed to fetch officers" });
-  }
-});
-
-router.post("/editOfficer", admin_authenticate, async (req, res) => {
-  const { id_number, name, position, email, course, year, campus } = req.body;
-
-  try {
-    const getAdmin = await Admin.findOne({
-      id_number: req.body.id_number,
-    });
-
-    const adminResult = await Admin.updateOne(
-      { id_number: id_number },
-      {
-        $set: {
-          name: name,
-          position: position,
-          campus: campus,
-          email: email,
-          course: course,
-          year: year,
-        },
-      }
-    );
-
-    if (adminResult.modifiedCount > 0) {
-      // Log the edit admin action
-      const log = new Log({
-        admin: req.user.name,
-        admin_id: req.user._id,
-        action: "Edited Admin",
-        target: `${id_number} - ${name}`,
-        target_id: getAdmin._id,
-        target_model: "Admin",
-      });
-
-      await log.save();
-
-      res.status(200).json({ message: "Officer updated successfully" });
-    } else {
-      res.status(404).json({ error: "No officer found with the provided ID" });
-    }
-  } catch (error) {
-    console.error("Error updating officer:", error);
-    res.status(500).json({ error: "Failed to update officer" });
-  }
-});
-
-router.post(
-  "/admin/change-password-officer",
-  admin_authenticate,
-  async (req, res) => {
-    try {
-      const getAdmin = await Admin.findOne({
-        id_number: req.body.id_number,
-      });
-
-      if (!getAdmin) {
-        return res.status(404).json({ message: "Admin not found" });
-      }
-
-      const hashedPassword = await bcrypt.hash(req.body.password, 10);
-      getAdmin.password = hashedPassword;
-      await getAdmin.save();
-
-      // Log the password change action
-      const log = new Log({
-        admin: req.user.name,
-        admin_id: req.user._id,
-        action: "Changed Admin Password",
-        target: `${getAdmin.id_number} - ${getAdmin.name}`,
-        target_id: getAdmin._id,
-        target_model: "Admin",
-      });
-
-      await log.save();
-
-      res.status(200).json({ message: "Password changed successfully" });
-    } catch (error) {
-      res
-        .status(500)
-        .json({ message: "An error occurred", error: error.message });
-    }
-  }
-);
-router.put("/admin/suspend", admin_authenticate, async (req, res) => {
-  const { id_number } = req.body;
-
-  try {
-    const updatedAdmin = await Admin.updateOne(
-      { id_number },
-      {
-        $set: {
-          status: "Suspend",
-        },
-      }
-    );
-
-    if (updatedAdmin.modifiedCount > 0) {
-      res.status(200).json({ message: "Admin status updated to Suspend" });
-    } else {
-      res.status(404).json({ message: "Admin not found or already suspended" });
-    }
-  } catch (error) {
-    console.error("Error suspending admin:", error);
-    res
-      .status(500)
-      .json({ message: "An error occurred", error: error.message });
-  }
-});
-
-//TODO: REMOVE ROLE FROM ADMIN
-router.put("/admin/role-remove", admin_authenticate, async (req, res) => {
-  const { id_number } = req.body;
-
-  try {
-    const updatedStudent = await Student.updateOne(
-      { id_number },
-      {
-        $set: {
-          role: "all",
-        },
-      }
-    );
-
-    const updatedStudentOrder = await Order.updateMany(
-      {
-        id_number,
-      },
-      {
-        $set: {
-          role: "all",
-        },
-      }
-    );
-
-    if (updatedStudent.modifiedCount > 0) {
-      res.status(200).json({ message: "Role removed successfully" });
-    } else {
-      res.status(404).json({ message: "Student not found" });
-    }
-  } catch (error) {
-    console.error("Error removing role from admin:", error);
-    res
-      .status(500)
-      .json({ message: "An error occurred", error: error.message });
-  }
-});
-
-router.put("/admin/restore-officer", admin_authenticate, async (req, res) => {
-  const { id_number } = req.body;
-
-  try {
-    const getAdmin = await Admin.findOne({
-      id_number: req.body.id_number,
-    });
-
-    const updatedAdmin = await Admin.updateOne(
-      { id_number },
-      {
-        $set: {
-          status: "Active",
-        },
-      }
-    );
-
-    if (updatedAdmin.modifiedCount > 0) {
-      // Log the restore officer action
-      const log = new Log({
-        admin: req.user.name,
-        admin_id: req.user._id,
-        action: "Restored Suspended Admin",
-        target: `${id_number} - ${getAdmin.name}`,
-        target_id: getAdmin._id,
-        target_model: "Admin",
-      });
-
-      await log.save();
-
-      res.status(200).json({ message: "Admin status updated to Active" });
-    } else {
-      res.status(404).json({ message: "Admin not found or already active" });
-    }
-  } catch (error) {
-    console.error("Error activating admin:", error);
-    res
-      .status(500)
-      .json({ message: "An error occurred", error: error.message });
-  }
-});
+//Dashboard Publish Merchandise Count
 router.get(
-  "/admin/search-student/:id_number",
+  "/merchandise-created",
   admin_authenticate,
-  async (req, res) => {
-    const { id_number } = req.params;
-
-    try {
-      const student = await Student.findOne({
-        id_number,
-        role: "all",
-      });
-      if (!student) {
-        res.status(404).json({
-          message: "Student not found or Student is already added a role",
-        });
-      } else {
-        res.status(200).json({ data: student });
-      }
-    } catch (error) {
-      console.error(error);
-      res
-        .status(500)
-        .json({ message: "An error occurred", error: error.message });
-    }
-  }
+  getPublishMerchandiseCountController
 );
-router.put("/admin/request-role", admin_authenticate, async (req, res) => {
-  const { id_number, role, admin } = req.body;
-
-  try {
-    const student = await Student.findOne({ id_number: id_number });
-    const updatedRole = await Student.updateOne(
-      { id_number },
-      {
-        $set: {
-          role: role,
-          isRequest: true,
-          adminRequest: admin,
-        },
-      }
-    );
-    const updatedStudentOrder = await Order.updateMany(
-      { id_number },
-      {
-        $set: {
-          role: role,
-        },
-      }
-    );
-    await new Log({
-      admin: admin,
-      action:
-        "Request Role for " + student.first_name + " " + student.last_name,
-      target: role + " request",
-      target_model: "Student",
-    }).save();
-
-    if (updatedRole.modifiedCount > 0) {
-      res.status(200).json({ message: "Role updated successfully" });
-    } else {
-      res.status(404).json({ message: "Student not found" });
-    }
-  } catch (error) {
-    console.error("Error updating student role:", error);
-    res
-      .status(500)
-      .json({ message: "An error occurred", error: error.message });
-  }
-});
-
-router.get("/admin/get-request-role", admin_authenticate, async (req, res) => {
-  try {
-    const students = await Student.find({ isRequest: true });
-    const users = students.map((student) => ({
-      id_number: student.id_number,
-      email: student.email,
-      name:
-        student.first_name +
-        " " +
-        student.middle_name +
-        " " +
-        student.last_name,
-      course: student.course,
-      year: student.year,
-      role: student.role,
-      isRequest: student.isRequest,
-      adminRequest: student.adminRequest,
-    }));
-    res.status(200).json({ data: users });
-  } catch (error) {
-    console.error("Error fetching students:", error);
-    res.status(500).json("Internal Server Error");
-  }
-});
-
-router.get("/admin/get-request-admin", admin_authenticate, async (req, res) => {
-  try {
-    const admin = await Admin.find({ status: "Request" });
-    const users = admin.map((admins) => ({
-      id_number: admins.id_number,
-      email: admins.email,
-      name: admins.name,
-      course: admins.course,
-      year: admins.year,
-      role: admins.position,
-      status: admins.status,
-    }));
-    res.status(200).json({ data: users });
-  } catch (error) {
-    console.error("Error fetching admins:", error);
-    res.status(500).json("Internal Server Error");
-  }
-});
-router.put("/admin/approve-role", admin_authenticate, async (req, res) => {
-  const { id_number } = req.body;
-
-  try {
-    const updatedRole = await Student.updateOne(
-      { id_number },
-      {
-        $set: {
-          isRequest: false,
-        },
-      }
-    );
-
-    if (updatedRole.modifiedCount > 0) {
-      res.status(200).json({ message: "Role approved successfully" });
-    } else {
-      res.status(404).json({ message: "Student not found" });
-    }
-  } catch (error) {
-    console.error("Error updating student role:", error);
-    res
-      .status(500)
-      .json({ message: "An error occurred", error: error.message });
-  }
-});
-
-router.put("/admin/decline-role", admin_authenticate, async (req, res) => {
-  const { id_number } = req.body;
-
-  try {
-    const updatedRole = await Student.updateOne(
-      { id_number },
-      {
-        $set: {
-          role: "all",
-          isRequest: false,
-        },
-      }
-    );
-
-    if (updatedRole.modifiedCount > 0) {
-      res.status(200).json({ message: "Role approved successfully" });
-    } else {
-      res.status(404).json({ message: "Student not found" });
-    }
-  } catch (error) {
-    console.error("Error updating student role:", error);
-    res
-      .status(500)
-      .json({ message: "An error occurred", error: error.message });
-  }
-});
-
-router.post("/admin/add-officer", admin_authenticate, async (req, res) => {
-  const {
-    id_number,
-    name,
-    password,
-    email,
-    position,
-    course,
-    year,
-    campus,
-    status,
-  } = req.body;
-
-  try {
-    const hashedPassword = await bcrypt.hash(password, 10);
-    const newAdmin = new Admin({
-      id_number,
-      name,
-      password: hashedPassword,
-      email,
-      position,
-      course,
-      year,
-      campus,
-      status,
-      access: "standard",
-    });
-    await newAdmin.save();
-
-    res.status(200).json({ message: "Account Creation successful" });
-  } catch (error) {
-    console.error(error);
-  }
-});
-
-//admin/approve-admin-account
-
-router.put(
-  "/admin/approve-admin-account",
+//Dashboard Placed Order Count
+router.get("/placed-orders", admin_authenticate, getOrderPlacedCountController);
+//All Student Dashboard Stats
+router.get(
+  "/dashboard-stats",
   admin_authenticate,
-  async (req, res) => {
-    const { id_number } = req.body;
-
-    try {
-      const updatedRole = await Admin.updateOne(
-        { id_number },
-        {
-          $set: {
-            status: "Active",
-          },
-        }
-      );
-
-      if (updatedRole.modifiedCount > 0) {
-        res
-          .status(200)
-          .json({ message: "Admin Account approved successfully" });
-      } else {
-        res.status(404).json({ message: "Admin not found" });
-      }
-    } catch (error) {
-      console.error("Error updating admin account:", error);
-      res
-        .status(500)
-        .json({ message: "An error occurred", error: error.message });
-    }
-  }
+  getStudentDashboardCountController
 );
-router.put(
-  "/admin/decline-admin-account",
+//Get Daily Sales Stats
+router.get("/get-daily-sales", admin_authenticate, getDailySalesController);
+//Get All admin officers
+router.get(
+  "/get-all-officers",
   admin_authenticate,
-  async (req, res) => {
-    const { id_number } = req.body;
-
-    try {
-      const deletedAdmin = await Admin.deleteOne({ id_number });
-
-      if (deletedAdmin.deletedCount > 0) {
-        res.status(200).json({ message: "Admin account deleted successfully" });
-      } else {
-        res.status(404).json({ message: "Admin not found" });
-      }
-    } catch (error) {
-      console.error("Error deleting admin account:", error);
-      res
-        .status(500)
-        .json({ message: "An error occurred", error: error.message });
-    }
-  }
+  getAllAdminAccountsController
+);
+//Get All Members
+router.get(
+  "/get-all-members",
+  admin_authenticate,
+  getAllAdminMembersController
+);
+//Get Suspend Admin Accounts
+router.get(
+  "/get-suspend-officers",
+  admin_authenticate,
+  getAllSuspendAdminAccountController
+);
+//Edit Admin Account
+router.post("/edit-officer", admin_authenticate, editAdminAccountController);
+//Change Password Admin Account
+router.post(
+  "/change-password-officer",
+  admin_authenticate,
+  changeAdminPasswordController
+);
+//Set Admin Account to Suspend
+router.put("/suspend", admin_authenticate, setSuspendAdminAccountController);
+//Set Member Role Remove
+router.put("/role-remove", admin_authenticate, setMemberRoleRemoveController);
+//Set Restore Deleted Admin
+router.put(
+  "/restore-officer",
+  admin_authenticate,
+  setRestoreAdminAccountController
+);
+//Set Admin Request Role to Member
+router.put("/request-role", admin_authenticate, setAdminRequestRoleController);
+//Get all Members Account Request Role
+router.get(
+  "/get-request-role",
+  admin_authenticate,
+  getAllRequestMemberController
+);
+//Get all Admin Account Request Role
+router.get(
+  "/get-request-admin",
+  admin_authenticate,
+  getAllRequestAdminAccountController
+);
+//Set Approve Role for Members
+router.put("/approve-role", admin_authenticate, approveRoleMemberController);
+//Set Decline Role for Members
+router.put("/decline-role", admin_authenticate, setDeclineMemberRoleController);
+//Add New Admin Account
+router.post("/add-officer", admin_authenticate, addNewAdminAccountController);
+//Approve Admin Account Creation
+router.put(
+  "/approve-admin-account",
+  admin_authenticate,
+  approveAdminAccountController
+);
+//Decline Admin Account Creation
+router.put(
+  "/decline-admin-account",
+  admin_authenticate,
+  declineAdminAccountController
 );
 //update-admin-access
 router.put(
-  "/admin/update-admin-access",
+  "/update-admin-access",
   admin_authenticate,
-  async (req, res) => {
-    const { id_number, newAccess } = req.body;
-
-    if (!id_number || !newAccess) {
-      return res
-        .status(400)
-        .json({ message: "id_number and newAccess are required" });
-    }
-
-    try {
-      const adminToUpdate = await Admin.findOne({ id_number });
-
-      if (!adminToUpdate) {
-        return res.status(404).json({ message: "Admin not found" });
-      }
-
-      adminToUpdate.access = newAccess;
-      await adminToUpdate.save();
-
-      await new Log({
-        admin: req.user.name,
-        action: `Change Access for ${adminToUpdate.name} to ${newAccess}`,
-        target: "Change Access",
-        target_model: "Admin",
-      }).save();
-
-      res.status(200).json({ message: "Access updated successfully" });
-    } catch (error) {
-      console.error("Error updating access account:", error);
-      res
-        .status(500)
-        .json({ message: "An error occurred", error: error.message });
-    }
-  }
+  setNewAdminAccessController
 );
 
 module.exports = router;
