+++ server-side/server.js
@@ -4,8 +4,7 @@ const cors = require("cors");
 const mongoose = require("mongoose");
 const app = express();
 const PORT = process.env.PORT || 5000;
-const registerRoutes = require("./routes/register");
-const loginRoutes = require("./routes/login");
+const indexRoutes = require("./routes/index");
 const adminRoutes = require("./routes/admin");
 const studentRoutes = require("./routes/students");
 const merchRoutes = require("./routes/merch");
@@ -34,10 +33,9 @@ mongoose
   .catch((err) => console.log(err));
 
 //Routes
-app.use("/api", registerRoutes);
-app.use("/api", loginRoutes);
+app.use("/api", indexRoutes);
 app.use("/api", studentRoutes);
-app.use("/api", adminRoutes);
+app.use("/api/admin", adminRoutes);
 app.use("/api/merch", merchRoutes);
 app.use("/api/orders", orderRoutes);
 app.use("/api/facebook", facebookRoutes);
